detection/requirements.txt
ultralytics
pydantic
wandb
pandas
pillow
pyyaml
